Enhance project-task HTML view: - Moved 'See Project' button below heading - Repositioned zone, region, multistate and status - Updated 'View LOI' button text to 'LOI PDF' - Added 'View Attach...' link in Task 3 (Agreement and Handover) to open custom_agreement path
